refactor: battery info bottom sheet refactor
